+++ R/Datasets.R
@@ -12,11 +12,24 @@
 #' 
 #' Data were collected on HG-U133_Plus_2, Affymetrix Human Genome U133 Plus 2.0
 #' Array.
-#' 
-#' @name micro_S
-#' @docType data
+#'
 #' @format A data frame with 54613 probesets measured 6 times throught 4 time
-#' points.
+#' points. It features 54613 rows and 24 numeric columns.
+#' Columns are named **exactly** as:
+#' \code{N1_S_T60}, \code{N1_S_T90}, \code{N1_S_T210}, \code{N1_S_T390},
+#' \code{N2_S_T60}, \code{N2_S_T90}, \code{N2_S_T210}, \code{N2_S_T390},
+#' \code{N3_S_T60}, \code{N3_S_T90}, \code{N3_S_T210}, \code{N3_S_T390},
+#' \code{N4_S_T60}, \code{N4_S_T90}, \code{N4_S_T210}, \code{N4_S_T390},
+#' \code{N5_S_T60}, \code{N5_S_T90}, \code{N5_S_T210}, \code{N5_S_T390},
+#' \code{N6_S_T60}, \code{N6_S_T90}, \code{N6_S_T210}, \code{N6_S_T390}.
+#'
+#' The ordering is strict: for each replicate \code{N1}, \code{N2}, \code{N3}, 
+#' \code{N4} , \code{N5}, {\code{N6}} the time points are 
+#' ordered \code{T60}, \code{T90}, \code{T210}, \code{T390}.
+#'
+#' @name micro_S
+#' @aliases micro_S
+#' @docType data
 #' @references Vallat, L., Kemper, C. A., Jung, N., Maumy-Bertrand, M.,
 #' Bertrand, F., \dots{}, Bahram, S. (2013). Reverse-engineering the genetic
 #' circuitry of a cancer cell with predicted intervention in chronic
@@ -28,9 +41,6 @@
 #' data(micro_S)
 #' 
 NULL
-
-
-
 
 
 #' Unstimulated control dataset
@@ -48,7 +58,22 @@ NULL
 #' Data were collected on HG-U133_Plus_2, Affymetrix Human Genome U133 Plus 2.0
 #' Array.
 #' 
+#' @format A data frame with 54613 probesets measured 6 times throught 4 time
+#' points. It features 54613 rows and 24 numeric columns.
+#' Columns are named **exactly** as:
+#' \code{N1_US_T60}, \code{N1_US_T90}, \code{N1_US_T210}, \code{N1_US_T390},
+#' \code{N2_US_T60}, \code{N2_US_T90}, \code{N2_US_T210}, \code{N2_US_T390},
+#' \code{N3_US_T60}, \code{N3_US_T90}, \code{N3_US_T210}, \code{N3_US_T390},
+#' \code{N4_US_T60}, \code{N4_US_T90}, \code{N4_US_T210}, \code{N4_US_T390},
+#' \code{N5_US_T60}, \code{N5_US_T90}, \code{N5_US_T210}, \code{N5_US_T390},
+#' \code{N6_US_T60}, \code{N6_US_T90}, \code{N6_US_T210}, \code{N6_US_T390}.
+#'
+#' The ordering is strict: for each replicate \code{N1}, \code{N2}, \code{N3}, 
+#' \code{N4} , \code{N5}, {\code{N6}} the time points are 
+#' ordered \code{T60}, \code{T90}, \code{T210}, \code{T390}.
+#' 
 #' @name micro_US
+#' @aliases micro_US
 #' @docType data
 #' @format A data frame with 54613 probesets measured 6 times throught 4 time
 #' points.