+++ tests/testthat/test-load.R
@@ -1,0 +1,3 @@
+test_that("Package 'CascadeData' loads quietly", {
+  expect_silent(library(CascadeData))
+})